+++ gnfuv-temp-exp6/send_temperature.py
@@ -16,7 +16,7 @@ DELTA = float(os.getenv('DELTA', 1))
 EXPERIMENT = float(os.getenv('EXP', 6))
 LOGDIR = str(os.getenv('LOGDIR', '/tmp'))
 
-WINDOWSIZE = float(os.getenv('WIND', 30))
+WINDOWSIZE = int(os.getenv('WIND', 30))
 THRESHOLD = float(os.getenv('THRESHOLD', 0.1))
 
 #variables needed
